+++ src/jsonCopy.ts
